Addressing review comments and adding more tests
