+++ hudi-common/src/main/java/org/apache/hudi/BaseHoodieTableFileIndex.java
@@ -49,7 +49,6 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
-import java.util.Set;
 import java.util.function.Function;
 import java.util.stream.Collectors;
 
@@ -165,10 +164,6 @@ public abstract class BaseHoodieTableFileIndex {
     return cachedAllInputFileSlices.entrySet()
         .stream()
         .collect(Collectors.toMap(e -> e.getKey().path, Map.Entry::getValue));
-  }
-
-  public Set<PartitionPath> getAllCachedPartitionPaths() {
-    return cachedAllInputFileSlices.keySet();
   }
 
   protected List<PartitionPath> getAllQueryPartitionPaths() {
